MEDIUM: session: protect sess conns list by idle_conns_lock
Introduce idle_conns_lock usage to protect manipulation to <priv_conns>
session members. This is the list of all connections attached to a
session.

Currently, this patch is unneeded as such sessions are only manipulated
on a single-thread. However, it will be required to trigger purging of
idle connections attached to a session from an external thread.