MEDIUM: session: account on server idle conns attached to session
This patch adds a new member <curr_sess_idle_conns> on the server. It
serves as a counter of idle connections attached on a session instead of
regular idle/safe trees. This is used only for private connections.

The objective is to provide an alternive method to detect if an idle
connections is referencing a server prior to its deletion. Currently,
this is not strictly necessary as idle connections are directly freed
via "delete server" handler. However, this cleanup will be replaced by
an asynchronous mechanism reusing connection purge list. Thus, this new
<curr_sess_idle_conns> counter will be useful here.

Careful: connections attached to a session but not idle will not be
accounted by this counter. These connections can still be detected via
srv_has_streams() so "delete server" will be safe.

Now, CO_FL_SESS_IDLE flag roles is extended : if present it directly
indicates the the connections is accounted in <curr_sess_idle_conns>
server counter. Note that session_unown_conn() won't decrement it. This
is handled via a new codepath in srv_release_conn(). The objective is to
ensure that the connection is still accounted if selected for purge but
not yet freed. This requires an adjustement on conn release path :
conn_backend_deinit() is now also called for idle private connections.