MEDIUM: server: reject idle conns if server in maintenance
Currently, when a server is set on maintenance mode, its idle connection
are purged. However, this does not prevent currently used connection to
become idle later on, even if the server is still off.

Change this behavior : an idle connection is now rejected by the server
if it is in maintenance. This is implemented with a new condition in
srv_add_to_idle_list() which returns an error value. This instructs
muxes to immediately free the connection in this case.

A similar change is also performed for each MUX I/O handler and in
conn_notify_mux(). If the connection is idle, first check if server is
in maintenance. If this is the case, the connection is not reinserted in
its list but instead immediately freed.