MINOR: muxes: adjust takeover with buf_wait interaction
Takeover operation defines an argument <release>. This is a boolean
which specify if connection resources which are deleted during the
operation must be reallocated.

Typically, it is set to false when the operation is performed to reuse a
connection. However, if takeover is used prior to be able to delete a
connection from another thread, <release> can be set to true.

Currently, this last possibility was only used on "delete server"
handler. This was used to remove any idle connections prior to remove a
server instance. As this operation was conducted under thread isolation,
connection could be removed from buf_wait if subscribed for a new
buffer. This operation is not thread-safe, as such it is considered only
valid under thread isolation, hence why it was only performed if
<release> is true. In the contrary case, takeover would return an error.

The previous patch removes takeover usage for idle connection cleanup
via "delete server". As such, there is no code path which set <release>
on takeover. However, takeover is also used to kill a connection when
applying strict-maxconn on a server. As described in the first
paragraph, it could be legitimate to set <release> to true in this case.
However, it is not performed under thread isolation, so it is first
necessary to remove the assumption on takeover implementation that if
<release> is true, thread isolation is active. This is the exact purpose
of this commit.

This patch thus simplifies takeover of every muxes. Now, if connection
is subscribed in buf_wait list, takeover fails, even if <release> is
true. This is necessary to be able to use takeover for connection
deletion without thread isolation.