MINOR: backend: set release on takeover for strict maxconn
If strict maxconn is active on a server, it may be necessary to kill an
idle connection to never exceed the configured value. This relies on
takeover, to be able to grab an idle connection from another thread and
free it immediately.

As takeover is used to delete a connection and not reuse it, it can set
<release> argument to true. This will remove extra allocations of
resources just freed so that the connection can migrate to a new thread.
As such, it is a small optimization for strict maxconn implementation.

Note that this patch depends on the previous one which removes any
assumption in takeover implementation that thread isolation is active if
<release> is true.