BUG/MAJOR: quic: fix padding of INITIAL packets
A QUIC datagram that contains an INITIAL packet must be padded to 1.200
bytes to prevent any deadlock due to anti-amplification protection. This
is implemented by encoding a PADDING frame on the last packet of the
datagram if necessary.

Previously, qc_prep_pkts() was responsible to activate padding when
calling qc_do_build_pkt() as it knows which packet is the last to
encode. However, this as the side-effect of preventing PING emission as
it was implemented in an else-if branch after padding. This was fixed by
the below commit by altering the logic, so that padding was set to false
when a PING was necessary. A dedicated block to reactivate padding was
present in qc_do_build_pkt() itself.

  217e467e89d15f3c22e11fe144458afbf718c8a8
  BUG/MINOR: quic: fix malformed probing packet building

However, the fix is incorrect and may generate packets with INITIAL
content not padded. This happens if the last QEL used to built a packet
is not the initial one, and probing is used with available frames. In
this case, no PING is emitted, but this bypassed the padding block
condition. This causes a BUG_ON() crash in qc_txb_store() which catches
these packets.

To properly fix this, revert to the original logic : qc_prep_pkts() is
responsible to activate INITIAL padding. To not interfere with PING
emission, qc_do_build_pkt() body is adjusted so that PING block is
moved up in the function and detached from the padding condition.

The main benefit from this patch is that INITIAL padding decision in
qc_prep_pkts() is clearer now.

Note that padding can also be activated by qc_do_build_pkt(), as packets
should be big enough for header protection decipher. However, this case
is different from INITIAL padding, so it is not covered by this patch.

This should be backported up to 2.6.