MINOR: session: document explicitely that session_add_conn() is safe
A set of recent patches have simplified management of backend connection
attached to sessions. The API is now strictier to prevent any misuse.

One of this change is the addition of a BUG_ON() in session_add_conn(),
which ensures that a connection is not attached to a session if its
<owner> field points to another entry.

This case could break due to NTLM as connection are turned as private
during transfers, possibly with several sessions already using it.
However, NTLM is now only applied for HTTP/1.1 as it does not make sense
if the connection is already shared. To clarify this, BUG_ON() comment
is extended.