MINOR: server: shard by thread sess_conns list storage in server struct
Server member <sess_conns> is a mt_list which stores every backend
connections attached to a session which targets this server. These
connecions are therefore not present in idle server trees.

Mostly, this contains idle/non-idle private connections. This can also
contain connections in use store temporarily in the session to prevent
sharing across clients, which is useful when backend multiplexer suffers
from head-of-line blocking.

The main utility of this list is to be able to cleanup idle private
connections prior to removing a server via "delete server" CLI. This
commit transforms <sess_conns> member, which is moved from server
structure to srv_per_thread. Effectively, the list is now shared by
threads.

There should be no functional change. However, this patch is necessary
to adapt connection removal for "delete server" to an asynchronous
method based on connections purging mechanism, as now it is easier to
loop over each private connections on a server one thread at a time.