MEDIUM: conn/muxes: reinsert BE priv conn into sess on IO completion
When IO handler of MUX is executed, special care must be taken if the
connection is considered as idle, as it could be stolen from another
thread via a takeover. Thus, connection is first removed from its idle
tree, and reinserted on completion, if the connection is still valid.

Idle private connections are not concerned by this, because takeover is
not applied on them. However, a future patch will implement purging of
these connections along with regular idle ones. As such, it is necessary
to also protect private connections usage now. This is the subject of
this patch and the next one.

On this patch, IO handlers ending of muxes is adjusted. If the
connection was idle and thus temporarily removed from its session on IO
handler start, a new code path is executed to reinsert the connection on
its session. This is implemented via a new function
session_reinsert_idle_conn(). Contrary to session_add_conn(), this is
reserved for reinserting idle connections after a temporary removal.

Also, contrary to _srv_add_idle() used by regular idle connections,
session_reinsert_idle_conn() may fail as an allocation is required. If
this happens, the connection is simply destroyed.

This patch is only a code adjustment, with no effect for now. Indeed,
session_reinsert_idle_conn() is protected by <conn_in_list> variable in
IO handlers, which is not yet set for private connections. Thus,
behavior changes will really start to occur with the next patch.