BUG/MINOR: server: decrement session idle_conns on delete server
When a server is deleted, each of its idle connections are removed. This
is also performed for every private connections stored on sessions which
referenced the target server.

All of these private connections are idle. This is guaranteed by
srv_check_for_deletion() which ensures that "delete server" handler is
only executed if there is no active connection on a server. A BUG_ON()
on CO_FL_SESS_IDLE is already present to enforce this.

Session is also responsible to maintain its counter for idle
connections. This is useful to limit the number of idle connections
usable per sessions. However, in case of "delete server", this counter
is not decremented. Fix this by ensuring that a decrement is performed
for every private connections removed.

This should be backported up to 3.0.